+++ NEWS.md
@@ -1,3 +1,5 @@
+# odbc (development version)
+
 # odbc 1.3.3
 
 * Hadley Wickham is now the maintainer.